requirements.txt
requests>=2.28.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
